# Use fused types in denoise, warp