--- a/skimage/restoration/_denoise.py
+++ b/skimage/restoration/_denoise.py
@@ -2,8 +2,8 @@ import scipy.stats
 import numpy as np
 from math import ceil
 from .. import img_as_float
-from ..restoration._denoise_cy import _denoise_bilateral, _denoise_tv_bregman, \
-    _gaussian_weight
+from ..restoration._denoise_cy import (_denoise_bilateral, _denoise_tv_bregman,
+                                       _gaussian_weight)
 from .._shared.utils import warn
 import pywt
 import skimage.color as color
@@ -156,8 +156,9 @@ def denoise_bilateral(image, win_size=None, sigma_color=None, sigma_spatial=1,
     # where needed within Cython.
     empty_dims = np.empty(dims, dtype=image.dtype)
 
-    return _denoise_bilateral(image, image.max(), win_size, sigma_color, sigma_spatial,
-                              bins, mode, cval, color_lut, range_lut, empty_dims, out)
+    return _denoise_bilateral(image, image.max(), win_size, sigma_color,
+                              sigma_spatial, bins, mode, cval, color_lut,
+                              range_lut, empty_dims, out)
 
 
 def denoise_tv_bregman(image, weight, max_iter=100, eps=1e-3, isotropic=True):