## Description
Fixes #3449, #1287 

Supersedes #3469, #3253 (I think), #1977. 

All credit to @AetherUnbound and @hmaarrfk here, I have just rebased their work on the latest master after @hmaarrfk fixed our negative indexing use in Cython.


## Checklist
- [x] Clean style in [the spirit of PEP8](https://www.python.org/dev/peps/pep-0008/)
- [x] [Docstrings for all functions](https://github.com/numpy/numpy/blob/master/doc/example.py)
- [x] ~~Gallery example in `./doc/examples` (new features only)~~
- [ ] Benchmark in `./benchmarks`, if your changes aren't covered by an
  existing benchmark
- [x] Unit tests

## For reviewers

- [ ] Check that the PR title is short, concise, and will make sense 1 year
  later.
- [ ] Check that new functions are imported in corresponding `__init__.py`.
- [ ] Check that new features, API changes, and deprecations are mentioned in
      `doc/release/release_dev.rst`.
- [ ] Consider backporting the PR with `@meeseeksdev backport to v0.14.x`