@pep8speaks 2018-10-19
Hello @jni! Thanks for submitting the PR.

 - In the file [`skimage/restoration/_denoise.py`](https://github.com/jni/scikit-image/blob/e0d7e5d560fc244ee292b5233b1dff0301fa7b65/skimage/restoration/_denoise.py), following                are the PEP8 issues :
> [Line 5:80](https://github.com/jni/scikit-image/blob/e0d7e5d560fc244ee292b5233b1dff0301fa7b65/skimage/restoration/_denoise.py#L5): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (80 > 79 characters)
> [Line 159:80](https://github.com/jni/scikit-image/blob/e0d7e5d560fc244ee292b5233b1dff0301fa7b65/skimage/restoration/_denoise.py#L159): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (87 > 79 characters)
> [Line 160:80](https://github.com/jni/scikit-image/blob/e0d7e5d560fc244ee292b5233b1dff0301fa7b65/skimage/restoration/_denoise.py#L160): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (86 > 79 characters)


 - There are no PEP8 issues in the file [`skimage/restoration/tests/test_denoise.py`](https://github.com/jni/scikit-image/blob/e0d7e5d560fc244ee292b5233b1dff0301fa7b65/skimage/restoration/tests/test_denoise.py) !