@hmaarrfk 2018-10-19
Why is OSX failing? https://travis-ci.org/scikit-image/scikit-image/jobs/443575582#L5300