@hmaarrfk 2018-10-19
@jni, I'm going to say that though I appreciate that my original PR is taken seriously, I don't know how worthwhile it is to complicate the warping code if it doesn't really give a speedup. In my experience, templated code is so much harder to read and to adapt. By including this, we are forcing future developers to also develop warping algorithms that function for all numeric types. This is a huge potential burden.

I would look into why it was that any speedup was obtained for float64 types and maybe just apply that fix. It might have been passing a parameter by reference in a tight loop which was necessary to template the functions.

I think the assumption that integer arithmetic is faster than floatingpoint is seriously flawed due to massive hardware optimizations for the floating point units in modern processors, though I'm not an expert in this field.

float32/float64 seems to be a valid assumption if you can guarantee not upcasting. float16 only recently started to exist because GPUs are running out of memory for machine learning.

I can't speak specifically for @AetherUnbound's PR, I was mostly helping with syntax.