@jni 2018-10-19
@hmaarrfk but speed is not the only advantage, there is also RAM usage, right? I think that's very valuable. And personally I don't find the code much more complicated. And you know I'm quite nitpicky about complicated code! =D

It looks like the test failure is to do with random noise values. Same failure on that Mac build as on AppVeyor:
https://travis-ci.org/scikit-image/scikit-image/jobs/443575582#L5285-L5301

Dunno why it would happen in this PR and not elsewhere, since that test appears to use float64 anyway... Any ideas?