@hmaarrfk 2018-10-19
Probably because we added some tests, the seed is different than before. Numpy (the repo) actually prints the values of the arrays that failed.

Would be useful if we had that....