@jni 2018-10-20
@hmaarrfk it's not that the seed is different, it's that it's not set. You can run

```
pytest skimage/restoration/tests/test_denoise.py::test_denoise_bilateral_2d
```

repeatedly to get different results each time.