@hmaarrfk 2018-10-20
@jni really? i thought the python seed was set to some consistent value everytime. I guess this isn't matlab again. It probably needs an environment variable.