@hmaarrfk 2018-10-20
```
export PYTHONHASHSEED=1   
pytest skimage/restoration/tests/test_denoise.py::test_denoise_bilateral_2d 
```
will fail, but I can't run it with `--pdb`