@hmaarrfk 2018-10-20
```
    def test_denoise_bilateral_2d():
        img = checkerboard_gray.copy()[:50, :50]
        # add some random noise
        noise = 0.5 * img.std() * np.random.rand(*img.shape)
        img += noise
        img = np.clip(img, 0, 1)
    
        out1 = restoration.denoise_bilateral(img, sigma_color=0.1,
                                             sigma_spatial=10, multichannel=False)
        out2 = restoration.denoise_bilateral(img, sigma_color=0.2,
                                             sigma_spatial=20, multichannel=False)
    
        # make sure noise is reduced in the checkerboard cells
>       assert img[30:45, 5:15].std() > out1[30:45, 5:15].std()
E       assert 0.07044972960234154 > nan
E        +  where 0.07044972960234154 = <built-in method std of numpy.ndarray object at 0x7fcbe9b3ce40>()
E        +    where <built-in method std of numpy.ndarray object at 0x7fcbe9b3ce40> = array([[ 0.12671524,  0.15149392,  0.21206663,  0.15423946,  0.00545805,\n         0.16724259,  0.21282153,  0.1807636 ...1418,  0.08659465,  0.017158  ,  0.18889751,\n         0.09346233,  0.08582789,  0.12415705,  0.2002917 ,  0.07256701]]).std
E        +  and   nan = <built-in method std of numpy.ndarray object at 0x7fcbe9b3cdf0>()
E        +    where <built-in method std of numpy.ndarray object at 0x7fcbe9b3cdf0> = array([[ nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan],\n       [ nan,  nan,  nan,  nan,  nan,  nan,  nan,...  nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan],\n       [ nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan,  nan]]).std
```