@hmaarrfk 2018-10-20
Do we have a benchmark? is it possible to see what happens when we change the looping order of the tight loop in `_denoise_tv_bregman` to `C` ordering?