@hmaarrfk 2018-10-20
Finally, when all those comments are addressed, you can release the gil. You will need this line of code somewhere

```
                      color_lut_bin = min(
                          <Py_ssize_t>(dist * dist_scale), max_color_lut_bin)
```