@hmaarrfk 2018-10-20
@jni, thanks for helping find the bug.

Can you please close this PR? it seems there is more work to do in the denoising PR.