@jni 2018-10-21
@hmaarrfk I don't see why the debugging needs to happens specifically in one or the other PR? I'm happy to take this on. Thanks for the detailed and specific review either way!