@jni 2018-11-29
@AetherUnbound I've given this another go. Sorry to miss your ping earlier this month. I think @hmaarrfk is on holiday in Japan atm. =D

Let's see how my latest commits fare in CI! =)