@jni 2018-11-29
One last thing I want to do is move all of the LUT code out of Cython. But for now it's bedtime.