@jni 2018-12-01
It may not look like it, but CI is actually happy! :stuck_out_tongue_closed_eyes: :tada: :tada: :tada: 

@scikit-image/core can we have some reviews please? =D This fixes a regression in 0.14.1 relative to 0.14.0 so it's relatively urgent (within scikit-image levels of urgency, anyway =P).

The failures are in `--pre` because, as far as I can tell, NumPy 1.16 changed the deprecation message for `numpy.matrix`, so all the work in #3242 is now obsolete. :man_facepalming: I'll raise a separate issue but I don't think it should hold up this PR. [edit: @sciunto already found it in #3571]

https://travis-ci.org/scikit-image/scikit-image/jobs/461638460#L3340

```
ValueError: Unexpected warning: the matrix subclass is not the recommended way to represent matrices or deal with linear algebra (see https://docs.scipy.org/doc/numpy/user/numpy-for-matlab-users.html). Please adjust your code to use regular ndarray.
```