@jni 2018-12-10
@scikit-image/core anyone want to take a stab at reviewing this?