@jni 2019-01-15
@scikit-image/core ping! As we are gearing up to release 0.14.2 any day now (ASAP given that we are broken with the newly release NumPy 1.16), this pull request would be very nice to have, since it fixes a bug introduced in 0.14.1.