@jni 2019-02-21
@hmaarrfk here are the results. master:

```
 $ asv dev -b WarpSuite
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Building for existing-py_home_jni_miniconda3_envs_cf_bin_python
[  0.00%] ·· Benchmarking existing-py_home_jni_miniconda3_envs_cf_bin_python
[ 50.00%] ··· Running benchmark_transform_warp.WarpSuite.time_same_type                                                     ok
[ 50.00%] ···· 
               =============== ====== ======== ======== ========
               --                               order           
               ---------------------- --------------------------
                   dtype_in      N       0        1        3    
               =============== ====== ======== ======== ========
                 numpy.uint8    128    132μs    357μs    1.22ms 
                 numpy.uint8    1024   12.5ms   27.4ms   63.9ms 
                 numpy.uint8    4096   314ms    566ms    1.20s  
                 numpy.uint16   128    139μs    372μs    933μs  
                 numpy.uint16   1024   12.5ms   29.4ms   66.6ms 
                 numpy.uint16   4096   323ms    577ms    1.25s  
                numpy.float32   128    134μs    381μs    1.20ms 
                numpy.float32   1024   15.3ms   35.7ms   74.8ms 
                numpy.float32   4096   355ms    686ms    1.20s  
                numpy.float64   128    128μs    479μs    1.05ms 
                numpy.float64   1024   17.9ms   35.7ms   72.5ms 
                numpy.float64   4096   317ms    565ms    1.14s  
               =============== ====== ======== ======== ========
```

PR HEAD:

```
 $ asv dev -b WarpSuite
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Building for existing-py_home_jni_miniconda3_envs_cf_bin_python
[  0.00%] ·· Benchmarking existing-py_home_jni_miniconda3_envs_cf_bin_python
[ 50.00%] ··· Running benchmark_transform_warp.WarpSuite.time_same_type                                                     ok
[ 50.00%] ···· 
               =============== ====== ======== ======== ========
               --                               order           
               ---------------------- --------------------------
                   dtype_in      N       0        1        3    
               =============== ====== ======== ======== ========
                 numpy.uint8    128    134μs    344μs    929μs  
                 numpy.uint8    1024   12.4ms   33.2ms   63.7ms 
                 numpy.uint8    4096   320ms    617ms    1.26s  
                 numpy.uint16   128    135μs    360μs    895μs  
                 numpy.uint16   1024   12.3ms   28.3ms   63.3ms 
                 numpy.uint16   4096   342ms    577ms    1.16s  
                numpy.float32   128    131μs    369μs    971μs  
                numpy.float32   1024   14.6ms   31.6ms   67.2ms 
                numpy.float32   4096   340ms    588ms    1.14s  
                numpy.float64   128    128μs    407μs    954μs  
                numpy.float64   1024   17.8ms   35.9ms   76.7ms 
                numpy.float64   4096   317ms    572ms    1.25s  
               =============== ====== ======== ======== ========
```
master:

```
[100.00%] ··· Running benchmark_transform_warp.WarpSuite.time_to_float64                                                    ok
[100.00%] ···· 
               =============== ====== ======== ======== ========
               --                               order           
               ---------------------- --------------------------
                   dtype_in      N       0        1        3    
               =============== ====== ======== ======== ========
                 numpy.uint8    128    129μs    363μs    931μs  
                 numpy.uint8    1024   11.6ms   26.2ms   61.4ms 
                 numpy.uint8    4096   311ms    537ms    1.10s  
                 numpy.uint16   128    122μs    387μs    1.00ms 
                 numpy.uint16   1024   11.4ms   25.7ms   66.6ms 
                 numpy.uint16   4096   306ms    535ms    1.12s  
                numpy.float32   128    139μs    385μs    997μs  
                numpy.float32   1024   14.3ms   31.4ms   64.6ms 
                numpy.float32   4096   302ms    570ms    1.12s  
                numpy.float64   128    125μs    377μs    967μs  
                numpy.float64   1024   17.7ms   33.6ms   70.2ms 
                numpy.float64   4096   310ms    567ms    1.14s  
               =============== ====== ======== ======== ========
```
PR HEAD:
```
[100.00%] ··· Running benchmark_transform_warp.WarpSuite.time_to_float64                                                    ok
[100.00%] ···· 
               =============== ====== ======== ======== ========
               --                               order           
               ---------------------- --------------------------
                   dtype_in      N       0        1        3    
               =============== ====== ======== ======== ========
                 numpy.uint8    128    143μs    380μs    990μs  
                 numpy.uint8    1024   11.7ms   27.1ms   63.3ms 
                 numpy.uint8    4096   305ms    534ms    1.13s  
                 numpy.uint16   128    128μs    368μs    1.02ms 
                 numpy.uint16   1024   11.5ms   27.1ms   63.8ms 
                 numpy.uint16   4096   308ms    526ms    1.12s  
                numpy.float32   128    142μs    379μs    1.34ms 
                numpy.float32   1024   14.0ms   29.8ms   69.3ms 
                numpy.float32   4096   302ms    551ms    1.18s  
                numpy.float64   128    118μs    355μs    926μs  
                numpy.float64   1024   18.0ms   35.0ms   72.2ms 
                numpy.float64   4096   308ms    561ms    1.17s  
               =============== ====== ======== ======== ========
```